requirements.txt
streamlit==1.28.1
pandas==2.1.3
requests==2.31.0
openpyxl==3.1.2
python-dotenv==1.0.0
plotly==5.17.0
numpy==1.24.3
seaborn==0.12.2
matplotlib==3.7.2
xlrd==2.0.1
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4.post3

